pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
